fix arrange windows & improve rust code - avoid memory leak

+++ src-tauri/src/data/window_info.rs
@@ -6,5 +6,5 @@ use serde::{Deserialize, Serialize};
 pub struct WindowInfo {
     pub frame: Frame,
     pub window_id: CGWindowID,
-    pub pid: i32,
+    pub pid: i64,
 }
